pyproject.toml
# pyproject.toml

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "codesum"
version = "0.1.1" # Incremented version
authors = [
  { name="Your Name", email="your.email@example.com" },
]
description = "Interactive code summarizer using AI and TUI"
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Environment :: Console :: Curses",
    "Topic :: Software Development :: Documentation",
    "Topic :: Utilities",
]
keywords = ["code", "summary", "ai", "openai", "tui", "curses", "documentation"]

dependencies = [
    "openai",
    "pathspec",
    "python-dotenv",
    "pyperclip",
    "platformdirs >= 4.0.0",
    "windows-curses; sys_platform == 'win32'",
    "importlib-resources; python_version < '3.9'", # Backport for older Python
]

[project.scripts]
codesum = "codesum.app:main"

# --- Add these sections for package data ---`
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
codesum = ["prompts/*.md"]
# --------------------------------------------